added user authorization to games controller

+++ apps/api/src/app/games/controllers/games.controller.ts
@@ -11,6 +11,7 @@ import {
   Post,
   Put,
   UseGuards,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { Review } from '../../../../../../shared/review';
 import { Game } from '../../../../../../shared/game';
@@ -70,10 +71,18 @@ export class GamesController {
     @Body() changes: Partial<Game>
   ): Promise<Game> {
     if (changes._id) {
-      throw new BadRequestException("Can't update course id");
+      throw new BadRequestException("Can't update game id");
     }
 
     const user = jwt.verify(authJwtToken, JWT_SECRET);
+    const verifyUser = await this.userDB.findUser(user.email);
+    const verifyGame = await this.gamesDB.findOne(gameId);
+    console.log(verifyUser._id);
+    console.log(verifyGame.userId);
+    if (verifyGame.userId.toString() !== verifyUser._id.toString()) {
+      throw new UnauthorizedException('Can not update other users games');
+    }
+
     await this.userDB.updateGame(user.email, gameId, changes);
 
     return this.gamesDB.updateGame(gameId, changes);
@@ -86,7 +95,12 @@ export class GamesController {
   ) {
     const user = jwt.verify(authJwtToken, JWT_SECRET);
     const game = await this.gamesDB.findOne(gameId);
-    game.reviews.forEach(review => {
+    const verifyUser = await this.userDB.findUser(user.email);
+    if (game.userId.toString() !== verifyUser._id.toString()) {
+      throw new UnauthorizedException('Can not delete other users games');
+    }
+
+    game.reviews.forEach((review) => {
       this.userDB.removeReviewById(review.userId, review._id);
     });
     await this.userDB.removeGame(user.email, gameId);
@@ -94,6 +108,14 @@ export class GamesController {
   }
 
   //Reviews
+  @Get('/review/:gameId/:reviewId')
+  async getReview(
+    @Param('reviewId') reviewId: string,
+    @Param('gameId') gameId: string
+  ): Promise<Review> {
+    return this.gamesDB.getReview(gameId, reviewId);
+  }
+
   @Put('/review/:gameId')
   async addReview(
     @Headers('authorization') authJwtToken,
@@ -120,6 +142,12 @@ export class GamesController {
     @Body() updatedReview: Review
   ): Promise<Game> {
     const user = jwt.verify(authJwtToken, JWT_SECRET);
+    const verifyUser = await this.userDB.findUser(user.email);
+    const verifyReview = await this.gamesDB.getReview(gameId, reviewId);
+    if (verifyReview.userId !== verifyUser._id) {
+      throw new UnauthorizedException('Can not update other users reviews');
+    }
+
     await this.userDB.updateReview(user.email, reviewId, updatedReview);
 
     await this.gamesDB.updateReview(gameId, reviewId, updatedReview);
@@ -133,6 +161,12 @@ export class GamesController {
     @Param('gameId') gameId: string
   ): Promise<Game> {
     const user = jwt.verify(authJwtToken, JWT_SECRET);
+    const verifyUser = await this.userDB.findUser(user.email);
+    const verifyReview = await this.gamesDB.getReview(gameId, reviewId);
+    if (verifyReview.userId !== verifyUser._id) {
+      throw new UnauthorizedException('Can not delete other users reviews');
+    }
+
     await this.userDB.removeReview(user.email, reviewId);
 
     await this.gamesDB.removeReview(gameId, reviewId);
@@ -142,26 +176,56 @@ export class GamesController {
   //Actors
   @Put('/actor/:gameId')
   async createActor(
+    @Headers('authorization') authJwtToken,
     @Param('gameId') gameId: string,
     @Body() actor: Actor
   ): Promise<Game> {
+    const user = jwt.verify(authJwtToken, JWT_SECRET);
+    const verifyGame = await this.gamesDB.findOne(gameId);
+    const verifyUser = await this.userDB.findUser(user.email);
+    if (verifyGame.userId.toString() !== verifyUser._id.toString()) {
+      throw new UnauthorizedException(
+        'Can not add actors to other users games'
+      );
+    }
+
     return this.gamesDB.createActor(gameId, actor);
   }
 
   @Put('/actor/:gameId/:actorId')
   async updateActor(
+    @Headers('authorization') authJwtToken,
     @Param('gameId') gameId: string,
     @Param('actorId') actorId: string,
     @Body() actor: Partial<Actor>
   ): Promise<Game> {
+    const user = jwt.verify(authJwtToken, JWT_SECRET);
+    const verifyGame = await this.gamesDB.findOne(gameId);
+    const verifyUser = await this.userDB.findUser(user.email);
+    if (verifyGame.userId.toString() !== verifyUser._id.toString()) {
+      throw new UnauthorizedException(
+        'Can not update actors from other users games'
+      );
+    }
+
     return this.gamesDB.updateActor(gameId, actorId, actor);
   }
 
   @Delete('/actor/:gameId/:actorId')
   async deleteActor(
+    @Headers('authorization') authJwtToken,
     @Param('gameId') gameId: string,
     @Param('actorId') actorId: string
   ): Promise<Game> {
+    const user = jwt.verify(authJwtToken, JWT_SECRET);
+    const verifyGame = await this.gamesDB.findOne(gameId);
+    const verifyUser = await this.userDB.findUser(user.email);
+    if (verifyGame.userId.toString() !== verifyUser._id.toString()) {
+      throw new UnauthorizedException(
+        'Can not delete actors from other users games'
+      );
+    }
+
     return this.gamesDB.deleteActor(gameId, actorId);
   }
 }
